+++ Bootstrap/Utils/CommandLine.cpp
@@ -80,8 +80,8 @@ void CommandLine::Read()
 				loadmode = 0;
 			InternalCalls::MelonHandler::LoadModeForMods = (InternalCalls::MelonHandler::LoadMode)loadmode;
 		}
-		else if (strstr(command, AddPrefixToLaunchOption("agfregenerate")))
-			AssemblyGenerator::ForceRegeneration = true;
+		//else if (strstr(command, AddPrefixToLaunchOption("agfregenerate")))
+		//	AssemblyGenerator::ForceRegeneration = true;
 		else if (strstr(command, AddPrefixToLaunchOption("agfvunity")))
 		{
 			std::string version = argv[i + 1];