+++ MelonLoader/Extensions/Layers/Harmony12/AccessTools.cs
@@ -3,12 +3,12 @@ using System.Reflection;
 
 namespace Harmony
 {
-    [Obsolete("Harmony.AccessTools is obsolete. Please use HarmonyLib.AccessTools instead.")]
+    [Obsolete("Harmony.AccessTools is Only Here for Compatibility Reasons. Please use HarmonyLib.AccessTools instead.")]
     public static class AccessTools
     {
-        [Obsolete("Harmony.AccessTools.Method is obsolete. Please use HarmonyLib.AccessTools.Method instead.")]
+        [Obsolete("Harmony.AccessTools.Method is Only Here for Compatibility Reasons. Please use HarmonyLib.AccessTools.Method instead.")]
         public static MethodInfo Method(Type type, string name, Type[] parameters = null, Type[] generics = null) => HarmonyLib.AccessTools.Method(type, name, parameters, generics);
-        [Obsolete("Harmony.AccessTools.Method is obsolete. Please use HarmonyLib.AccessTools.Method instead.")]
+        [Obsolete("Harmony.AccessTools.Method is Only Here for Compatibility Reasons. Please use HarmonyLib.AccessTools.Method instead.")]
         public static MethodInfo Method(string typeColonMethodname, Type[] parameters = null, Type[] generics = null) => HarmonyLib.AccessTools.Method(typeColonMethodname, parameters, generics);
     }
 }