+++ MelonLoader/Properties/AssemblyInfo.cs
@@ -10,4 +10,4 @@ using System.Runtime.InteropServices;
 [assembly: Guid("A662769A-B294-434F-83B5-176FC4795334")]
 [assembly: AssemblyVersion(MelonLoader.BuildInfo.Version)]
 [assembly: AssemblyFileVersion(MelonLoader.BuildInfo.Version)]
-[assembly: MelonLoader.HarmonyShield]
+[assembly: MelonLoader.PatchShield]