+++ SM_Il2Cpp/Properties/AssemblyInfo.cs
@@ -10,4 +10,4 @@ using System.Runtime.InteropServices;
 [assembly: Guid("CD98B694-D833-41B0-AF57-A6B9D024B045")]
 [assembly: AssemblyVersion(MelonLoader.BuildInfo.Version)]
 [assembly: AssemblyFileVersion(MelonLoader.BuildInfo.Version)]
-[assembly: MelonLoader.HarmonyShield]
+[assembly: MelonLoader.PatchShield]